+++ code/pipeline-rust/src/config/policy.rs
@@ -14,9 +14,9 @@ pub enum Policy {
         /// into the consumer buffer.
         base_nanos: u64,
 
-        /// The maximum wait time in nanoseconds before giving up and dropping
-        /// the data.
-        max_nanos: u64,
+        /// The maximum accumulated wait time in nanoseconds before giving up
+        /// and dropping the data.
+        max_nanos: f64,
 
         /// The multiplier to apply to the wait time after each retry. For
         /// example, a multiplier of 2.0 will double the wait time after each
